Feedback scenegraph & test sound
Reorganized the scenegraph. The scene now contains a root game object and adds objects as children to the root. The gameobjects now update their children instead of the scene updating all objects. Changed from shared ptrs to components and children to unique for the owner and raw ptr for others.

Added a test for sound.